+++ packages/component/src/createStyleSheet.js
@@ -212,7 +212,7 @@ const createScrollBarStyle = ({ autoHide, scrollBarHeight, scrollBarMargin }) =>
 });
 
 const createRootStyle = ({ autoHide, autoHideFlipperOnEdge }) => ({
-  '& .react-film__root__content': {
+  '&.react-film__root .react-film__root__content': {
     overflow: 'hidden',
     position: 'relative'
   },
@@ -226,36 +226,26 @@ const createRootStyle = ({ autoHide, autoHideFlipperOnEdge }) => ({
               transitionDelay: '0s'
             },
 
-            '& .react-film__flipper .react-film__flipper__slider': {
-              transitionDelay: '0s'
-            },
+            '& .react-film__flipper': {
+              '& .react-film__flipper__slider': {
+                transitionDelay: '0s'
+              },
 
-            '& .react-film__flipper.react-film__flipper--left': {
-              ...(autoHideFlipperOnEdge
-                ? {
-                    '&:not(.react-film__flipper--hide) .react-film__flipper__slider': {
-                      left: 0
-                    }
-                  }
-                : {
-                    '& .react-film__flipper__slider': {
-                      left: 0
-                    }
-                  })
-            },
+              '&.react-film__flipper--left': {
+                [autoHideFlipperOnEdge
+                  ? '&:not(.react-film__flipper--hide) .react-film__flipper__slider'
+                  : '& .react-film__flipper__slider']: {
+                  left: 0
+                }
+              },
 
-            '& .react-film__flipper.react-film__flipper--right': {
-              ...(autoHideFlipperOnEdge
-                ? {
-                    '&:not(.react-film__flipper--hide) .react-film__flipper__slider': {
-                      right: 0
-                    }
-                  }
-                : {
-                    '& .react-film__flipper__slider': {
-                      right: 0
-                    }
-                  })
+              '&.react-film__flipper--right': {
+                [autoHideFlipperOnEdge
+                  ? '&:not(.react-film__flipper--hide) .react-film__flipper__slider'
+                  : '& .react-film__flipper__slider']: {
+                  right: 0
+                }
+              }
             }
           }
         }